modulerized components/chat
