+++ frontend/src/components/chat/components/sidebar/Sidebar.tsx
@@ -1,0 +1,69 @@
+import type { ExpandedSections, Document, ChatHistory } from '../../types'
+import { SidebarHeader } from './SidebarHeader'
+import { SourcesSection } from './SourcesSection'
+import { ChatHistorySection } from './ChatHistorySection'
+import { KnowledgeBaseSection } from './KnowledgeBaseSection'
+
+interface SidebarProps {
+  expandedSections: ExpandedSections
+  toggleSection: (section: keyof ExpandedSections) => void
+  selectedMessageSources: any[]
+  chatHistory: ChatHistory[]
+  selectedDocuments: number[]
+  documents: Document[]
+  onShowDocumentModal: () => void
+  onRemoveDocument: (docId: number) => void
+}
+
+export function Sidebar({
+  expandedSections,
+  toggleSection,
+  selectedMessageSources,
+  chatHistory,
+  selectedDocuments,
+  documents,
+  onShowDocumentModal,
+  onRemoveDocument
+}: SidebarProps) {
+  return (
+    <div 
+      data-sidebar="true"
+      style={{ 
+        width: '350px', 
+        minWidth: '350px', 
+        flexShrink: 0, 
+        backgroundColor: '#1f2937', 
+        color: 'white', 
+        height: '100vh',
+        overflowY: 'auto',
+        overflowX: 'hidden'
+      }} 
+      className="flex flex-col"
+    >
+      <SidebarHeader />
+      
+      <div style={{ flex: 1, overflowY: 'auto', overflowX: 'hidden', display: 'flex', flexDirection: 'column' }}>
+        <SourcesSection 
+          expandedSections={expandedSections}
+          toggleSection={toggleSection}
+          selectedMessageSources={selectedMessageSources}
+        />
+        
+        <ChatHistorySection 
+          expandedSections={expandedSections}
+          toggleSection={toggleSection}
+          chatHistory={chatHistory}
+        />
+        
+        <KnowledgeBaseSection 
+          expandedSections={expandedSections}
+          toggleSection={toggleSection}
+          selectedDocuments={selectedDocuments}
+          documents={documents}
+          onShowDocumentModal={onShowDocumentModal}
+          onRemoveDocument={onRemoveDocument}
+        />
+      </div>
+    </div>
+  )
+}
