+++ frontend/src/components/chat/components/sidebar/SidebarHeader.tsx
@@ -1,0 +1,8 @@
+export function SidebarHeader() {
+  return (
+    <div style={{ borderBottom: '1px solid #4b5563', padding: '16px' }}>
+      <h2 style={{ fontSize: '18px', fontWeight: '600', color: 'white', margin: 0 }}>Assistant Panel</h2>
+      <p style={{ fontSize: '12px', color: '#d1d5db', margin: '4px 0 0 0' }}>Document insights and controls</p>
+    </div>
+  )
+}
